examples/start-agents/google_adk_starter/requirements.txt
google-adk>=0.8.0
google-genai>=0.1.0
python-dotenv>=1.1.0
streamlit>=1.32.0